+++ packages/components/src/Footer.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 export default function Footer() {
   return (
-    <footer>
+    <footer className="bg-good-dog-celadon">
       <Image
         className="relative"
         src="/bg-assets/footer_checker.svg"
@@ -11,36 +11,45 @@ export default function Footer() {
         height={0}
         style={{ width: "100%", height: "auto" }}
       />
-      <div className="flex flex-row justify-center space-x-4 bg-good-dog-celadon pb-16 pt-36">
+      <div className="flex flex-row items-center justify-between pb-16 pt-36">
+        <div className="flex flex-grow flex-row justify-center space-x-4">
+          <Image
+            src="/icons/facebook_icon.svg"
+            width={70}
+            height={70}
+            alt="good-dog-facebook"
+          />
+          <Image
+            src="/icons/twitter_icon.svg"
+            width={70}
+            height={70}
+            alt="good-dog-twitter"
+          />
+          <Image
+            src="/icons/instagram_icon.svg"
+            width={70}
+            height={70}
+            alt="good-dog-instagram"
+          />
+          <Image
+            src="/icons/email_icon.svg"
+            width={70}
+            height={70}
+            alt="good-dog-email"
+          />
+          <Image
+            src="/icons/youtube_icon.svg"
+            width={70}
+            height={70}
+            alt="good-dog-youtube"
+          />
+        </div>
         <Image
-          src="/icons/facebook_icon.svg"
-          width={70}
-          height={70}
-          alt="good-dog-facebook"
-        />
-        <Image
-          src="/icons/twitter_icon.svg"
-          width={70}
-          height={70}
-          alt="good-dog-twitter"
-        />
-        <Image
-          src="/icons/instagram_icon.svg"
-          width={70}
-          height={70}
-          alt="good-dog-instagram"
-        />
-        <Image
-          src="/icons/email_icon.svg"
-          width={70}
-          height={70}
-          alt="good-dog-email"
-        />
-        <Image
-          src="/icons/youtube_icon.svg"
-          width={70}
-          height={70}
-          alt="good-dog-youtube"
+          className="ml-auto mr-12"
+          src="/icons/Dark Mode Logo.svg"
+          width={120}
+          height={120}
+          alt="good-dog-logo"
         />
       </div>
     </footer>
